move store_image related classes to services and refactor them
