--- app/jobs/store_image/object_key_builder.rb
@@ -1,19 +1,0 @@
-require "securerandom"
-
-module StoreImage
-  class ObjectKeyBuilder
-    include Memery
-
-    def initialize(filename:)
-      @filename = filename
-    end
-
-    memoize def object_key
-      "images/#{Time.now.utc.strftime('%Y%m%d')}/#{SecureRandom.uuid}-#{filename}"
-    end
-
-    private
-
-    attr_reader :filename
-  end
-end
