+++ app/services/store_image/upload/content_type_resolver.rb
@@ -1,0 +1,26 @@
+module StoreImage
+  module Upload
+    class ContentTypeResolver
+      include Memery
+
+      def initialize(filename:)
+        @filename = filename
+      end
+
+      memoize def content_type
+        return "image/png" if extension == ".png"
+        return "image/jpeg" if [".jpg", ".jpeg"].include?(extension)
+
+        raise ArgumentError, "Unsupported image extension: #{extension}"
+      end
+
+      private
+
+      attr_reader :filename
+
+      def extension
+        File.extname(filename).downcase
+      end
+    end
+  end
+end
